Format logs as JSON
